requirements.test.txt
pre-commit
pytest
pytest-cov
pytest-homeassistant-custom-component
# From manifest
iolite-client==0.5.1
